wip(permissions): implementing new structure
