+++ src/interface/graphql/adapters/authorization/resolvers/guarded-connection.resolver.ts
@@ -1,4 +1,4 @@
-import { Resolver } from '@nestjs/graphql'
+import { Parent, ResolveField, Resolver } from '@nestjs/graphql'
 
 import { AccessControl } from '@adapters/authorization/access-control.adapter'
 import { QueryGuardAdapter } from '@adapters/authorization/query-guard.adapter'
@@ -9,6 +9,9 @@ import { CoreEntityInterface } from '@core/core-entity.interface'
 import { CoreEntity } from '@core/core.orm-entity'
 import { CoreProvider } from '@core/core.provider'
 import { CoreEntityProvider } from '@core/entity.provider'
+import { GuardedNodeGraphQLInterface } from '@interface/graphql/adapters/authorization/interfaces/guarded-node.interface'
+import { ConnectionPolicyGraphQLObject } from '@interface/graphql/adapters/authorization/objects/connection-policy.object'
+import { RequestUserWithContext } from '@interface/graphql/adapters/context/decorators/request-user-with-context.decorator'
 import { BaseGraphQLResolver } from '@interface/graphql/resolvers/base.resolver'
 
 import { GuardedConnectionGraphQLInterface } from '../interfaces/guarded-connection.interface'
@@ -30,6 +33,14 @@ export abstract class GuardedConnectionGraphQLResolver<
     super()
 
     this.queryGuard = new QueryGuardAdapter(resource, core, coreEntityProvider)
+  }
+
+  @ResolveField('policy', () => ConnectionPolicyGraphQLObject)
+  protected async resolveNodePolicy(
+    @Parent() node: GuardedNodeGraphQLInterface,
+    @RequestUserWithContext() userWithContext: UserWithContext,
+  ) {
+    return this.getConnectionPolicyForUserWithArgs(userWithContext, node.id)
   }
 
   protected async getConnectionPolicyForUserWithArgs(
