+++ src/interface/graphql/modules/key-result/access-control/key-result.access-control.ts
@@ -1,20 +1,19 @@
 import { Injectable } from '@nestjs/common'
 
-import { AccessControl } from '@adapters/authorization/access-control.adapter'
 import { Command } from '@adapters/policy/enums/command.enum'
 import { Resource } from '@adapters/policy/enums/resource.enum'
 import { UserWithContext } from '@adapters/state/interfaces/user.interface'
 import { KeyResult } from '@core/modules/key-result/key-result.orm-entity'
 import { Team } from '@core/modules/team/team.orm-entity'
-import { User } from '@core/modules/user/user.orm-entity'
 import { CorePortsProvider } from '@core/ports/ports.provider'
+import { KeyResultBaseAccessControl } from '@interface/graphql/modules/key-result/access-control/base.access-control'
 
 @Injectable()
-export class KeyResultAccessControl extends AccessControl {
+export class KeyResultAccessControl extends KeyResultBaseAccessControl {
   protected readonly resource = Resource.KEY_RESULT
 
-  constructor(private readonly core: CorePortsProvider) {
-    super()
+  constructor(protected core: CorePortsProvider) {
+    super(core)
   }
 
   public async canCreate(_user: UserWithContext, _id: string): Promise<boolean> {
@@ -31,37 +30,12 @@ export class KeyResultAccessControl extends AccessControl {
 
     const isKeyResultOwner = await this.isKeyResultOwner(keyResult, user)
     const isTeamLeader = await this.isTeamLeader(teams, user)
-    const isCompanyMember = await this.isCompanyMember(keyResult, user)
+    const isCompanyMember = await this.isKeyResultCompanyMember(keyResult, user)
 
     return this.canActivate(user, Command.UPDATE, isKeyResultOwner, isTeamLeader, isCompanyMember)
   }
 
   public async canDelete(_user: UserWithContext, _id: string): Promise<boolean> {
     return false
-  }
-
-  private async isKeyResultOwner(keyResult: KeyResult, user: UserWithContext): Promise<boolean> {
-    const owner = await this.core.dispatchCommand<User>('get-key-result-owner', keyResult)
-
-    return owner.id === user.id
-  }
-
-  private async isTeamLeader(teams: Team[], user: UserWithContext): Promise<boolean> {
-    const leaderPromises = teams.map(async (team) =>
-      this.core.dispatchCommand<User>('get-team-owner', team),
-    )
-    const leaders = await Promise.all(leaderPromises)
-
-    return leaders.some((leader) => leader.id === user.id)
-  }
-
-  private async isCompanyMember(keyResult: KeyResult, user: UserWithContext): Promise<boolean> {
-    const keyResultCompany = await this.core.dispatchCommand<Team>(
-      'get-key-result-company',
-      keyResult,
-    )
-    const userCompanies = await this.core.dispatchCommand<Team[]>('get-user-companies', user)
-
-    return userCompanies.some((company) => company.id === keyResultCompany.id)
   }
 }
