+++ src/interface/graphql/modules/key-result/key-result.module.ts
@@ -3,13 +3,13 @@ import { Module } from '@nestjs/common'
 import { GraphQLConfigModule } from '@config/graphql/graphql.module'
 import { CoreModule } from '@core/core.module'
 import { AmplitudeModule } from '@infrastructure/amplitude/amplitude.module'
-import { KeyResultAccessControl } from '@interface/graphql/modules/key-result/key-result.access-control'
+import { KeyResultCheckInAccessControl } from '@interface/graphql/modules/key-result/access-control/key-result-check-in.access-control'
+import { KeyResultCommentAccessControl } from '@interface/graphql/modules/key-result/access-control/key-result-comment.access-control'
+import { KeyResultAccessControl } from '@interface/graphql/modules/key-result/access-control/key-result.access-control'
 
 import { KeyResultCheckInsConnectionGraphQLResolver } from './check-in/connections/key-result-check-ins/key-result-check-ins.resolver'
-import { KeyResultCheckInAccessControl } from './check-in/key-result-check-in.access-control'
 import { KeyResultCheckInGraphQLResolver } from './check-in/key-result-check-in.resolver'
 import { KeyResultCommentsConnectionGraphQLResolver } from './comment/connections/key-result-comments/key-result-comments.resolver'
-import { KeyResultCommentAccessControl } from './comment/key-result-comment.access-control'
 import { KeyResultCommentGraphQLResolver } from './comment/key-result-comment.resolver'
 import { KeyResultKeyResultCheckInsConnectionGraphQLResolver } from './connections/key-result-key-result-check-ins/key-result-key-result-check-ins.resolver'
 import { KeyResultKeyResultCommentsConnectionGraphQLResolver } from './connections/key-result-key-result-comments/key-result-key-result-comments.resolver'
@@ -24,14 +24,14 @@ import { KeyResultGraphQLResolver } from './key-result.resolver'
     KeyResultCommentGraphQLResolver,
     KeyResultCheckInGraphQLResolver,
     KeyResultCheckInsConnectionGraphQLResolver,
-    KeyResultCheckInAccessControl,
-    KeyResultCommentAccessControl,
     KeyResultsConnectionGraphQLResolver,
     KeyResultCommentsConnectionGraphQLResolver,
     KeyResultKeyResultCommentsConnectionGraphQLResolver,
     KeyResultKeyResultCheckInsConnectionGraphQLResolver,
     KeyResultTimelineConnectionGraphQLResolver,
     KeyResultAccessControl,
+    KeyResultCheckInAccessControl,
+    KeyResultCommentAccessControl,
   ],
 })
 export class KeyResultGraphQLModule {}
